Pin some crate versions explicitly
We have had a few issues with the flatgeobuf and flatbuffer
crates drifting incompatibly across minor releases, so I
think it may be safer to hard-pin these to an exact version.

The issue we have seen is that by default when installing a binary
crate from crates.io (e.g. cargo install geoq), it does not honor
the Cargo.lock file included with the repo.

This means that cargo will re-resolve the compatible version graph
within the scope of what you have specified in your Cargo.toml
file. And it turns out this can in some cases result in conflicts,
e.g. if a dep we depend on upgrades and introduces a conflict with
something else we depend on.